Replace custom proxy with goproxy library (#45)
- Use github.com/elazarl/goproxy for HTTP proxy and CONNECT handling
- ConnectHijack preserves CONNECT-level skill auth for MITM'd requests
- Unified processRequest() handles HTTP, MITM HTTPS, and transparent TLS
- Split proxy_test.go into domain-specific test files (connect, helm, packages, registry)
- Fix MITM test: write 200 OK in hijack func (goproxy doesn't for ConnectHijack)
- Fix test buffering: pass CONNECT request to http.ReadResponse to avoid body read hang
